feat : 코드 컨벤션에 맞춘 개행 수정, 주문 엔티티 필드 totalPrice 타입 변경

+++ src/main/java/dev/nbcsparta/assignment/commerce_backoffice/service/OrderService.java
@@ -55,7 +55,7 @@ public class OrderService {
         // 재고 차감 진행
         product.buyProduct(request.quantity());
 
-        long totalPrice = (long) product.getPrice() * request.quantity();
+        int totalPrice = product.getPrice() * request.quantity();
 
         Order order = new Order(
                 request.quantity(),
@@ -85,7 +85,6 @@ public class OrderService {
 
     @Transactional(readOnly = true)
     public OrderDetail getDetailOrder(Long orderId) {
-
         Order order = validateOrder(orderId);
 
         return OrderDetail.from(order);
@@ -100,11 +99,8 @@ public class OrderService {
 
     @Transactional
     public void delete(Long orderId) {
-
         Order order = validateOrder(orderId);
-
         int canceledQuantity = order.getQuantity();
-
         order.getProduct().cancelProduct(canceledQuantity);
 
         orderRepository.deleteById(orderId);
